+++ src/Components/Runner/Adapters/Docker/DockerImageCommandRunner.php
@@ -76,12 +76,7 @@ class DockerImageCommandRunner implements TestRunnerInterface
         foreach ($this->envVariables as $env) {
             $envCommands .= ' --env ' . $env . ' ';
         }
-
-        # first pull in quiet mode
-        # then we don't have such a long cli output
-        $cmd = "docker pull -q " . $this->dockerImage;
-        $this->shellRunner->execute($cmd);
-
+        
         $cmd = "docker run --rm " . $envCommands . " " . $this->dockerImage . " bash -c '" . $command . "'";
 
         if ($this->debugMode) {